Migrate relay-specific schema classes
