+++ saas/rti/schemas.py
@@ -138,21 +138,3 @@ class ProcessorStatus(BaseModel):
     state: str = Field(..., title="State", description="The state of the processor.", example="initialised")
     pending: List[Job] = Field(..., title="Pending Jobs", description="A list of pending jobs that are queued for execution.")
     active: List[Job] = Field(..., title="Active Jobs", description="A list of active jobs that are currently being executed by the processor (if any).")
-
-
-class Permission(BaseModel):
-    """
-    A permission, created by a data object owner, and needed by an RTI to decrypt a data object.
-    """
-    req_id: str = Field(..., title="Request Id", description="The id of the request (generated by the RTI).")
-    content_key: str = Field(..., title="Protected Content Key", description="The content key and some meta information encrypted by the a ephemeral key provided by the requesting RTI.")
-
-
-class DeployParameters(BaseModel):
-    """
-    Parameters needed for deploying a processor.
-    """
-    deployment: Union[Literal['native', 'docker']] = Field(..., title="Deployment", description="Indicates whether a processor should be deployed natively or using docker.")
-    encrypted_ssh_credentials: Optional[str] = Field(title="SSH Credentials", description="SSH credentials need by the RTI to deploy the processor on a remote machine. The credentials have to be encrypted using the public encryption key of the RTI node.")
-    encrypted_github_credentials: Optional[str] = Field(title="Github Credentials", description="Github credentials need by the RTI to clone the repository that contains the processor. The credentials have to be encrypted using the public encryption key of the RTI node.")
-    gpp_custodian: Optional[str] = Field(title="GPP Custodian", description="Information about which nodes hosts the GPP information for this processor. The RTI will use this information to retrieve the GPP for this processor.")
